// ConcentradorWeb/src/app/models/producto-financiero.ts
export class ProductoFinanciero {
    idConsultaEntidadProducto: number;
    idEntidadProducto: number;
    pocentajeTea: number;
    porcentajeComision: number;
    porcentajeSeguro: number;
    plazoMinimoMes: number;
    PlazoMaximoMes: number;
    descripcion: string;
    mensajeOpcional: string;
    detalleProducto: string;
    resumenProducto: string;
    plazoMinimoDia: number;
    PlazoMaximoDia: number;
    tieneFondoSeguro: boolean;
    nombreEntidadFinanciera: string;
    imagenEntidadFinanciera: string;
    TipoInstitucionId: number;
    DepartamentoId: number;
    codigoProductoFinanciero: string;
    nombreProductoFinanciero: string;
    descripcionProductoFinanciero: string;
    idDetalleEntidadProducto: number;
    TipoMonedaId: number;
    tipoMoneda: string;
    IngresoPermitido: number;
    MontoMinimoAceptable: number;
    MontoMaximoAceptable: number;
    montoMinimoDeposito: number;
    MontoMaximoDeposito: number;
}